+++ samples/simple_tenant_app.py
@@ -11,7 +11,7 @@ from dotenv import load_dotenv
 from inputimeout import inputimeout, TimeoutOccurred
 
 from c8y_api.app import SimpleCumulocityApp
-from c8y_api.model import Celsius, Device, Measurement
+from c8y_api.model import Celsius, Device, Measurement, Operation
 
 # A simple (per tenant) Cumulocity application can be created just like this.
 # The authentication information is read from the standard Cumulocity
@@ -37,7 +37,8 @@ for d in c8y.device_inventory.get_all(page_size=100):
     print(f"  {d.name} #{d.id}")
 
 # Creating devices
-new_device = Device(c8y, type='test_SomeDevice', name='MyTestDevice', custom_fragment={'foo': 'bar'}).create()
+new_device = Device(c8y, type='test_SomeDevice', name='MyTestDevice', custom_fragment={'foo': 'bar'},
+                    com_cumulocity_model_Agent={}).create()
 print(f"\nCreated new device: {new_device.name} #{new_device.id}")
 
 # Creating Measurements
@@ -46,6 +47,18 @@ for v in range(0, 10):
     m = Measurement(c8y, type='test_SomeMeasurementType', source=new_device.id,
                     c8y_TemperatureMeasurement={'t': Celsius(v)}).create()
     print(f"  Created measurement: #{m.id}, JSON: {m.to_full_json()}")
+
+# Creating Operation
+print("\nOperation")
+new_operation = Operation(c8y, new_device.id, 'Shell command', c8y_Command={'text': 'myCommand'})
+new_operation.create()
+
+operationList = c8y.operations.get_all(agent_id=new_device.id, status='PENDING', page_size=1)
+pending_operation = operationList[0]
+print(pending_operation.status)
+
+pending_operation.status = 'EXECUTING'
+pending_operation.update()
 
 
 # Cleaning up